chore: Formatting and move eip 7930 type
